pyproject.toml
[tool.poetry]
name = "gamesight"
version = "0.1.0"
description = "Gameplay video analysis pipeline for game studios."
authors = ["Research.tech"]
readme = "README.md"
packages = [{ include = "gamesight", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
aiosqlite = "^0.22.0"
fastapi = "^0.135.0"
ffmpeg-python = "^0.2.0"
google-genai = "^1.68.0"
pydantic = "^2.12.0"
pydantic-settings = "^2.11.0"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
yt-dlp = ">=2026.3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.2.0"
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.basedpyright]
reportMissingImports = "warning"
reportMissingModuleSource = "warning"
